api_traductor_api/requirements.txt
Pillow==8.1.0
pytesseract
googletrans==3.1.0a0